Conflicts:

	puluo_app_backend/core/src/main/java/com/puluo/dao/impl/PuluoUserDaoImpl.java
	puluo_app_backend/core/src/main/java/com/puluo/entity/PuluoUser.java

	puluo_app_backend/core/src/main/java/com/puluo/entity/impl/PuluoUserImpl.java

+++ puluo_app_backend/core/src/main/java/com/puluo/entity/PuluoEventInfo.java
@@ -15,7 +15,7 @@ public interface PuluoEventInfo {
 	int level(); 
 	int type(); 
 	int likes();
-	float rating(); 	
+	double rating(); 	
 	List<PuluoEventPoster> poster();
 	PuluoEventPoster thumbnail();
